requirements.txt
streamlit==1.36.0
chromadb==0.5.5
pypdf==5.0.0
sentence-transformers==3.0.1
numpy==1.26.4
google-generativeai==0.4.0